ams: fix MS_GetPartialFile/MS_GetPartialBody remaining-param LP64 truncation (int* vs callers' long*), plus sibling MS_ProcessNewMessages FirstError mismatch
FossilOrigin-Name: 729ddc5f0d4a09fd65c8f49af044b22f980950b5c3d4c1c0e4327496ad8b0d16

+++ src/ams/libs/ms/getbody.c
@@ -34,11 +34,11 @@ static char rcsid[]="$Header: /afs/cs.cmu.edu/project/atk-dist/auis-6.3/ams/libs
 #include <andrewos.h>
 #include <ms.h>
 extern int ConsiderLoggingRead(char *FileName);
-extern int MS_GetPartialFile(char *FileName, char *Buf, int BufLim, int offset, int *remaining, int *ct);
+extern int MS_GetPartialFile(char *FileName, char *Buf, int BufLim, int offset, long *remaining, int *ct);
 extern int QuickGetBodyFileName(char *DirName, char *id, char *FileName);
 extern int ReadOrFindMSDir(char *Name, struct MS_Directory **pDir, int Code);
 
-int MS_GetPartialBody(char *DirName, char *id, char *Buf, int BufLim, int offset, int *remaining, int *ct)
+int MS_GetPartialBody(char *DirName, char *id, char *Buf, int BufLim, int offset, long *remaining, int *ct)
 {
     struct MS_Directory *Dir;
     char FileName[MAXPATHLEN+1];
